requirements.txt
jinja2
typeguard
pyserial==3.5
pyyaml